# internal/app/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Go Connect API
  description: Service marketplace backend API (Phase 1–2).
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /admin/categories:
    post:
      tags: [Categories]
      summary: Create category
      operationId: createCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategorySuccessEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from login or register
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: Validation failed
            error: VALIDATION_ERROR
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: Email already registered
            error: CONFLICT
...
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: Authentication required
            error: UNAUTHORIZED
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: Account is not active
            error: FORBIDDEN
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: Something went wrong
            error: INTERNAL_ERROR
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          nullable: true
    ErrorEnvelope:
      type: object
      required: [success, message, error]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
...
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
          example: Plumbing
        description:
          type: string
          nullable: true
          example: Home plumbing and repairs
        is_active:
          type: boolean
          default: true
    CategorySuccessEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/CategoryResponse"
...